+++ Controls/_interface/TPaddingSize.ts
@@ -1,0 +1,13 @@
+/**
+ * @typedef {String} Controls/_interface/TPaddingSize
+ * @description Допустимые значения для отступов.
+ * @variant XS
+ * @variant S
+ * @variant M
+ * @variant L
+ * @variant XL
+ * @variant XXL
+ * @variant null
+ * todo XXL некорректное значение https://online.sbis.ru/opendoc.html?guid=6625e538-36f0-465a-af2e-4b4d1465b739
+ */
+export type TPaddingSize = 'XS'|'S'|'M'|'L'|'XL'|'XXL'|'null'|'default';
