+++ Controls/_treeTile/display/TreeTileCollectionItem.ts
@@ -6,12 +6,12 @@ import {
     TActionMode,
     TGradientType,
     TImagePosition,
-    TContentPadding,
     TImageViewMode,
     TTitlePosition,
     TShadowVisibility,
     TTitleStyle
 } from 'Controls/tile';
+import {TPaddingSize} from 'Controls/interface';
 import {ITreeItemOptions, TreeItem} from 'Controls/display';
 import { TemplateFunction } from 'UI/Base';
 import * as FolderIcon from 'wml!Controls/_treeTile/render/FolderIcon';
@@ -252,7 +252,7 @@ export default class TreeTileCollectionItem<T extends Model = Model>
         gradientType: TGradientType = 'dark',
         titleStyle: TTitleStyle = 'light',
         imagePosition: TImagePosition = 'top',
-        contentPadding: TContentPadding = 'default'
+        contentPadding: TPaddingSize = 'default'
     ): string {
         let classes = super.getTitleWrapperClasses(itemType, titleLines, gradientType, titleStyle,
             imagePosition, contentPadding);
