+++ Controls/interface.ts
@@ -87,3 +87,4 @@ export {IInputPlaceholder, IInputPlaceholderOptions} from './_interface/IInputPl
 export {IInputTag, IInputTagOptions} from './_interface/IInputTag';
 export {default as ITextValue} from './_interface/ITextValue';
 export {default as ISelectField, ISelectFieldsOptions} from './_interface/ISelectFields';
+export {TPaddingSize} from 'Controls/_interface/TPaddingSize';
